Aggiornamento della searchbar sostituzione della searchbar di giovanni
