--- src/components/searchbar/searchbar1.scss
@@ -1,37 +1,0 @@
-.box{
-    background: var(--secondary-color);
-    flex-direction: row;
-    height: 40px;
-    width: 100%;
-    border-radius: 1%;
-
-    .bottom-groups{
-        margin-bottom: 15px;
-        color: var(--foreground-color);
-        background: none;
-        
-        display: flex;
-        flex-direction: row;
-        margin-left: auto;
-        justify-content: flex-end; /* Aggiunto per posizionare a destra */
-   
-
-        .bott{
-            margin-bottom: 15px;
-            margin-right: 35px;
-            padding: 10px;
-            border-radius: 50px;
-            width: 15%;
-            height: 10%;
-            background: var(--neutral-color-dark);
-            display: flex;
-        }
-    
-    }
-}
-@media (width <= 1200px) {
-
-    .bottom-groups {
-        display: none !important;
-    }
-}
